+++ nrutil.h
@@ -57,7 +57,7 @@ int **imatrix(long nrl, long nrh, long ncl, long nch);
 float **submatrix(float **a, long oldrl, long oldrh, long oldcl, long oldch,
 	long newrl, long newcl);
 float **convert_matrix(float *a, long nrl, long nrh, long ncl, long nch);
-float ***f3tensor(long nrl, long nrh, long ncl, long nch, long ndl, long ndh);
+double ***f3tensor(long nrl, long nrh, long ncl, long nch, long ndl, long ndh);
 void free_vector(float *v, long nl, long nh);
 void free_ivector(int *v, long nl, long nh);
 void free_cvector(unsigned char *v, long nl, long nh);
@@ -68,7 +68,7 @@ void free_dmatrix(double **m, long nrl, long nrh, long ncl, long nch);
 void free_imatrix(int **m, long nrl, long nrh, long ncl, long nch);
 void free_submatrix(float **b, long nrl, long nrh, long ncl, long nch);
 void free_convert_matrix(float **b, long nrl, long nrh, long ncl, long nch);
-void free_f3tensor(float ***t, long nrl, long nrh, long ncl, long nch,
+void free_f3tensor(double ***t, long nrl, long nrh, long ncl, long nch,
 	long ndl, long ndh);
 
 #else /* ANSI */
@@ -79,7 +79,7 @@ float *vector();
 float **matrix();
 float **submatrix();
 float **convert_matrix();
-float ***f3tensor();
+double ***f3tensor();
 double *dvector();
 double **dmatrix();
 int *ivector();
